Revert "Reland "Run R8 on bouncycastle with shrinking for the bo..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/12942ec1-dd03-4993-8784-d4d0a6438084). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4cef000e80144eeec6cc149bf03e0ec4e2bbe907
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com