Revert "Completely revert R8 on bouncycastle."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/64587b2e-7d00-4dca-a8e7-c85d8976c91e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib8a9458e4ce991aa96cc255f87c0868bc0912e2a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com