Revert "Completely revert R8 on bouncycastle."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/79615c12-7a09-4c64-81e9-d29a836347fd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I29723effdfcfa79d5277aa16e4b44d45d3af6c34
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com