Revert "Revert "Revert "Revert "Revert "Merge remote-tracking br..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5c91cfee-9423-44da-8f45-b04678f85424). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 394639963
Change-Id: I981926ba416baa8c3a3b6a5ac0ea0a25befb6a0d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com