Revert "Revert^6 Merge remote-tracking branch 'aosp/upstream-mas..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3a4b01d6-1f4a-4552-9bd1-ec4a703370de). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 395924161
Change-Id: I1fd548260301ba0a212861fa047237e1806948d2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com