Revert "Revert^6 Merge remote-tracking branch 'aosp/upstream-mas..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0b520e9c-7ed2-4af3-89c7-7731e379ee92). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I135f8bf5ca40afa6331e2ee907cbfda05f236165
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com