Revert "Revert^6 Merge remote-tracking branch 'aosp/upstream-mas..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/20a1523d-1b17-4ec4-b667-905c03f46090). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1406054c6ff6194c8725bfbeb87a9023298fbf8a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com