Revert "Revert^6 Merge remote-tracking branch 'aosp/upstream-mas..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a44803e3-b108-4c9b-b3e3-80c138cc94b4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 397158861
Change-Id: Ic1d3cc44589a9828da7fc68da7077261d9fb10a1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com