Revert "Revert^6 Merge remote-tracking branch 'aosp/upstream-mas..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/73bc5b66-9c62-4df2-8a0a-80c81c5cf280). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie11d990a96e489d4e8cf1cdd7bcf7eb08b8af248
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com